chore: move productionrow to types
